+++ hadoop-testutils/src/main/java/org/apache/hadoop/hbase/oss/Hadoop33EmbeddedS3ClientFactory.java
@@ -20,13 +20,18 @@ package org.apache.hadoop.hbase.oss;
 import java.io.IOException;
 import java.net.URI;
 
-import org.apache.hadoop.fs.s3a.S3ClientFactory;
-import org.apache.hadoop.hbase.oss.EmbeddedS3.EmbeddedAmazonS3;
-
 import com.amazonaws.auth.AWSCredentialsProvider;
 import com.amazonaws.services.s3.AmazonS3;
 
-public class Hadoop32EmbeddedS3ClientFactory implements S3ClientFactory {
+import org.apache.hadoop.fs.s3a.S3ClientFactory;
+import org.apache.hadoop.hbase.oss.EmbeddedS3.EmbeddedAmazonS3;
+
+/**
+ * An S3ClientFactory for Hadoop 3.3 releases which have the change from
+ * HADOOP-13551. Builds on top of Hadoop32EmbeddedS3ClientFactory.
+ */
+public class Hadoop33EmbeddedS3ClientFactory implements S3ClientFactory {
+
   public AmazonS3 createS3Client(URI name) {
     AmazonS3 s3 = new EmbeddedAmazonS3();
     s3.createBucket(EmbeddedS3.BUCKET);
@@ -46,5 +51,13 @@ public class Hadoop32EmbeddedS3ClientFactory implements S3ClientFactory {
       String bucket,
       AWSCredentialsProvider credentialSet) {
     return createS3Client(name);
+  }
+
+  public AmazonS3 createS3Client(URI uri,
+      S3ClientCreationParameters s3ClientCreationParameters)
+      throws IOException {
+    AmazonS3 s3 = new EmbeddedAmazonS3();
+    s3.createBucket(uri.getHost());
+    return s3;
   }
 }
