+++ hbase-oss/README.md
@@ -136,26 +136,23 @@ src/test/resources/auth-keys.xml, which should be ignored by source control.
 
 ### Hadoop Versions
 
-HBoss mainly depends on *org.apache.hadoop.fs.FileSystem* contract, and
-current HBoss version is compatible with Hadoop releases *3.2.2* and *3.3.1*.
+HBoss is compiled against Hadoop 3.3.2; this allows it to implement
+hadoop 3.3-only APIs and provide the behavior which the contract
+tests expect.
 
-There are Maven profiles defined for the above mentioned Hadoop 3 versions. 
-Support for Hadoop 2 has been dropped as off Oct 2021.
-These are activated via the property `hadoop.profile`. These profiles choose
-a specific Hadoop release in that major line, defaulting to versions as defined
-in `hadoop32.version` and `hadoop33.version`. By default, Hadoop 3.3 is used by
-the build.
+The hadoop-3.3 profile exists and can still be enabled, however it does
+not change the build in any way.
+
 
 ### HBase Versions
 HBoss testing suite relies on HBase Zookeeper testing utility class, which has
 been changing over different HBase releases. Current HBoss version is guaranteed
 to compile with HBase *2.3.6*. Support for different HBase versions, 
-including HBase 1 through profiles has been dropped as off Oct 2021.
+including HBase 1 through profiles has been dropped as of Oct 2021.
 
 ### Examples of different build profile usage/combinations, and related dependency versions picked:
 
-    mvn verify                                      # Defaults to Hadoop 3.3.1, HBase 2.3.6
-    mvn verify -Dhadoop.profile=3.2                 # Activates Hadoop 3.2.2, HBase 2.3.6
-    mvn verify -Dhadoop.profile=3.3                 # Activates Hadoop 3.3.1, HBase 2.3.6
+    mvn verify                                      # Defaults to Hadoop 3.3.2, HBase 2.3.6
+    mvn verify -Dhadoop.profile=3.3                 # no difference
